Adds mobile app development skills
Adds Flutter as a skill to the home page and experiences section.
This change enhances the representation of skills on the home page
to include mobile app development. Updates the experiences section
with an "application programmer" skill entry.

Adds mobile app development skills

Adds Flutter as a skill to the home page and experiences section.

Enhances the representation of skills on the home page to include mobile app development. Updates the experiences section with an "application programmer" skill entry.